requirements.txt
langchain
langgraph
python-dotenv
pydantic
langchain-google-genai
google-generativeai
fastapi
uvicorn